Modify NiceObjectDef's `init` arg to take strings
The original implementation took a function. This was meant to allow the
user to pass a wrapped function to use as the initializer. In practice
this didn't work because the function's aren't created by the time the
init arg is processed, they still exist as signature tuples. Using
strings also allows you to use functions that may be defined later.