Allow specifying units that `len` signature uses
Sometimes a buffer/array length that's passed to a C function must be
specified in bytes rather than elements. This new syntax lets you use
'len:lentype', where `lentype` is the name of a C type. This specifies
that the buffer is len*sizeof(buftype)//sizeof(lentype) bytes long. The
special string `byte` can be used to measure in bytes, to guard against
the unlikely event that sizeof(char) != 1.

What may be a bit confusing is that if you use e.g. 'len=4:byte', that
still means your array is 4 *elements* long, but the value passed to the
C function is in terms of bytes. So, 'len=4:byte' for an array of 32-bit
ints will create an `int[4]` and pass a value of `16` to the function.