Add a better way of manually wrapping functions
Sometimes you just have to manually write a mid-level wrapper for a
library function, usually because the interface is convoluted and
unusual. However, it may still be very useful to do auto-wrapping so you
don't have to handle the nitty-gritty. This change allows you to access
the auto-wrapped version of the function from within your custom
wrapper.

After defining the function, give it a `sig` attribute which is the
usual sig tuple that you'd use to define an auto-wrapper. If you do
this, NiceLib will automatically attach an `orig` attribute which refers
to the auto-wrapped function.

This feature is currently experimental and only supported for "methods"
of NiceObjects.