Rename `ret_wrap` to `ret`
What used to be called return value wrappers are more aptly called
return value handlers. To reflect this change, I'm getting rid of the
old name of `ret_wrap` in favor of the existing shorthand `ret`. This
has the added benefit of consistency, since now there's no special
shorthand.

The old name will currently work for backwards compatibility, but will
produce a warning that the code should be update. I'll probably remove
support for the old name in the 0.4 release.

This change is a bit of a pain in the short-term, but worth it in the
long term.