Fix handling of lib paths
Now that we find and use the full path to each library, we have to be
careful about how we generate Python source from that. The old code just
put the contents of the path string between single-quotes. This has the
unfortunate issue that backslashes do not get escaped, resulting in
escape chars instead of a proper path.

This fix directly uses the string's repr instead to auto-handle the
escaping. I believe it also could've been solved by making the generated
code use a raw string instead.