Change error for missing functions into a warning
When developing mid-level wrappers it's essential that typos are
immediately and easily discovered. Originally this meant that a function
definition which didn't match any known functions would produce an
error. However, this has the unfortunate limitation of causing errors
when a user is using an older version of a library that may not have
some new functions. So, I've changed this to a warning instead for now.

This is still kinda annoying for those users, so maybe there's a better
solution which may involve having a separate 'dev' mode or something,
though that sounds overcomplicated.