Overhaul internals as part of introducing ArgHandlers
Moved much of the NiceLib subclass-creation code into NiceMeta.__init__.
This way we can split the logic into individual methods without having
to pass so many arguments around. Instead, we just use instance
attributes.

Tried to come up with a sensible division of responsibilities between
NiceMeta, Sig, ArgHandler, and LibFunction. This may continue to change
for a bit, but seems to be in a pretty good state.

Introduced LibMethods, which act as methods of NiceObjects. They are
implemented in part by turning LibFunctions into descriptors. This
helped us to simplify the NiceObject class-creation logic.

I'm sure I'm forgetting something, but that's most of it.