Add hybrid-func support to NiceObjects and change its implementation
The original way of attaching the auto-wrapped LibFunction as an
attribute of the hybrid function worked for functions within the NiceLib
subclass, but posed problems for those within NiceObject subclasses. In
NiceObjects, the hybrid-func (a function/unbound method) would
automatically get bound by Python's machinery, but its libfunc argument
is left unbound. There's no simple way to deal with this; you'd probably
have to make your own custom descriptor for hybrid functions, with the
added code and complications that that would bring.

Instead, we attach the auto-wrapped libfunc as a specially-named
attribute of the NiceLib/NiceObject subclass, so it gets automatically
bound like the rest of the LibFunctions.

I've also changed hybrid-funcs defined in a NiceLib to classmethods, to
give easier access to this attribute.