Support libs with absolute paths and change to lib dir when loading
You can give a lib's full path to dlopen, but that lib's directory isn't
added to the search path, so loading dependent libs from that same
directory can fail. Here we chdir to the directory where the header is
found, since Windows checks this directory for DLLs. Not sure about
Linux, we may need a separate solution (like setting LD_LIBRARY_PATH) if
this doesn't work.