Add custom pycparser extension for parsing C++isms
This should make it easier to handle C++isms in some cases, like
pass-by-reference syntax. The initial implementation treats cases of
pass-by-reference as pointers and ignores type-annotated enums,
generating an ordinary pycparser AST.

We can extend the AST class if it ever becomes useful, but ignoring the
special syntax should be good enough in many cases.