Add building of tables in development mode installs
This will generate lextab.py and yacctab.py in the source directory when
you run a development install. This will prevent generation of these
files in whatever directory you happen to be in when you import a
development-mode version of nicelib.

Note that if you want regeneration of these files, you'll have to run
a reinstall command. It's conceivable that we might arrange
auto-generation of these files in the proper directory when in
development mode, but I haven't looked into this.